pyproject.toml
[project]
name = "common-japanese-words-in-yoasobi-lyrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pytest>=8.3.5",
    "cutlet>=0.5.0",
    "SQLAlchemy>=2.0.40",
    "pandas>=2.2.3",
    "pytest-mock>=3.14.0",
    "psycopg2-binary>=2.9.10",
    "sudachipy>=0.6.10",
    "sudachidict-core>=20250129",
    "unidic>=1.1.0",
    "seaborn>=0.13.2",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.6",
    "japanize-matplotlib>=1.1.3",
]
